+++ Sources/StreamChat/APIClient/Endpoints/ChannelEndpoints.swift
@@ -103,11 +103,12 @@ extension Endpoint {
         )
     }
 
-    static func sendMessage(cid: ChannelId, messagePayload: MessageRequestBody, skipPush: Bool)
+    static func sendMessage(cid: ChannelId, messagePayload: MessageRequestBody, skipPush: Bool, skipEnrichUrl: Bool)
         -> Endpoint<MessagePayload.Boxed> {
         let body: [String: AnyEncodable] = [
             "message": AnyEncodable(messagePayload),
-            "skip_push": AnyEncodable(skipPush)
+            "skip_push": AnyEncodable(skipPush),
+            "skip_enrich_url": AnyEncodable(skipEnrichUrl)
         ]
         return .init(
             path: .sendMessage(cid),