+++ Sources/StreamChat/Controllers/ChannelController/ChannelController.swift
@@ -900,6 +900,7 @@ public extension ChatChannelController {
     ///     and `ChatMessageAttachmentSeed`s.
     ///   - quotedMessageId: An id of the message new message quotes. (inline reply)
     ///   - skipPush: If true, skips sending push notification to channel members.
+    ///   - skipEnrichUrl: If true, skips url enriching.
     ///   - extraData: Additional extra data of the message object.
     ///   - completion: Called when saving the message to the local DB finishes.
     ///
@@ -911,6 +912,7 @@ public extension ChatChannelController {
         mentionedUserIds: [UserId] = [],
         quotedMessageId: MessageId? = nil,
         skipPush: Bool = false,
+        skipEnrichUrl: Bool = false,
         extraData: [String: RawJSON] = [:],
         completion: ((Result<MessageId, Error>) -> Void)? = nil
     ) {
@@ -936,6 +938,7 @@ public extension ChatChannelController {
             mentionedUserIds: mentionedUserIds,
             quotedMessageId: quotedMessageId,
             skipPush: skipPush,
+            skipEnrichUrl: skipEnrichUrl,
             extraData: extraData
         ) { result in
             self.callback {